Added logic to put songs to playlists

+++ ui/components/PlaylistTrack/PlaylistTrack.jsx
@@ -1,14 +1,25 @@
-import React from 'react'
+import React, { Component } from "react";
 import './PlaylistTrack.css'
 
-export const PlaylistTrack = ({song, index}) => (
-    <div className="playlist-track">
-        <div className="playlist-track-title" title={song.title}>
-            <span className="number">{index + 1}</span>
-            {song.title}
-        </div>
-        <button type="button" className="delete-from-playlist" title="Delete from Playlist">
-            <span className="material-icons">delete</span>
-        </button>
-    </div>
-)
+export class PlaylistTrack extends Component {
+
+    render() {
+        let deleteButton;
+        if (!this.props.fromModal) {
+            deleteButton = (
+                <button type="button" className="delete-from-playlist" title="Delete from Playlist">
+                    <span className="material-icons">delete</span>
+                </button>
+            )
+        }
+        return (
+            <div className="playlist-track">
+                <div className="playlist-track-title" title={this.props.song.title}>
+                    <span className="number">{this.props.index + 1}</span>
+                    {this.props.song.title}
+                </div>
+                {deleteButton}
+            </div>
+        );
+    }
+}
